Update modal is now a separate component
